Add player accounts, game pacing, session weather, and Docker/Caddy deployment
Fixes from health check:
- Fix /api/race/start 500: numpy.bool from forecast averaging was not JSON
  serializable; cast forecast values to plain Python types
- Serve unpaid steward penalties at race finish instead of dropping them
- Wire /api/weather/current and /api/weather/forecast to the live race
  session instead of returning canned dry weather (forecast previously 404ed)
- Fix TypeScript build errors (fontsource imports, ring color style,
  nullable classification entry, unreachable ?? operand)

Gameplay:
- Slow the SSE race stream to strategy-friendly pacing (1x = 10s/lap,
  up to 20x = 0.5s/lap) and show seconds-per-lap in the dashboard HUD

Gamification:
- Basic accounts (register/login, PBKDF2 hashing, server-side tokens)
- Race results validated against the server-side engine and scored with
  F1 championship points; career stats, tiers, history, global standings
- New Driver Profile page; results auto-recorded from the Results page

Deployment:
- Dockerfiles for backend (FastAPI) and frontend (Vite build + Caddy SPA
  server proxying /api), docker-compose joining the shared reverse-proxy
  network, served at https://pitwall.34.17.100.46.sslip.io via sslip.io

+++ project/backend/app/simulation/weather.py
@@ -365,12 +365,14 @@ class WeatherSystem:
             most_common = max(set(conditions), key=conditions.count)
 
             # Average numerical values across all simulations
-            avg_air_temp = np.mean([path[lap_idx].air_temp for path in all_paths])
-            avg_track_temp = np.mean([path[lap_idx].track_temp for path in all_paths])
-            avg_humidity = np.mean([path[lap_idx].humidity for path in all_paths])
-            avg_rain = np.mean([path[lap_idx].rain_intensity for path in all_paths])
-            avg_dampness = np.mean([path[lap_idx].track_dampness for path in all_paths])
-            avg_wind = np.mean([path[lap_idx].wind_speed for path in all_paths])
+            # Cast to plain float — np.float64 propagates np.bool_ into API
+            # responses (FastAPI's encoder cannot serialize numpy bools).
+            avg_air_temp = float(np.mean([path[lap_idx].air_temp for path in all_paths]))
+            avg_track_temp = float(np.mean([path[lap_idx].track_temp for path in all_paths]))
+            avg_humidity = float(np.mean([path[lap_idx].humidity for path in all_paths]))
+            avg_rain = float(np.mean([path[lap_idx].rain_intensity for path in all_paths]))
+            avg_dampness = float(np.mean([path[lap_idx].track_dampness for path in all_paths]))
+            avg_wind = float(np.mean([path[lap_idx].wind_speed for path in all_paths]))
 
             forecast.append(WeatherState(
                 condition=most_common,
